// angular/src/app/core/models/maintenance-operation/maintenance-calling.model.ts
import { BaseModel } from '../base.model';
import { GlobalValidator } from '../../../shared/form-validation/validators';
import { Validators } from '@angular/forms';

export interface MaintenanceCallingModel extends BaseModel {
  licensePlates?: string;
  customertype?: string;
  expecteddate?: string;
  isContactSuccessful?: string;
  updateKm?: string;
  forgetKm?: string;
  nextBD?: string;
  dateTime?: string;
  carSale?: string;
  transferWork?: string;
  errorCVDV?: string;
  customersComplained?: string;
  callContent?: string;
  checkCall?: string;
  errorContent?: string;
  errorVin?: string;
  sortAgency?: string;
  errorModel?: string;
  carType?: string;
  carColor?: string;
  carOrigin?: string;
  carPrice?: string;
  carPurchase?: string;
  carPriceto?: string;
  yearManufacture?: string;
  timePurchase?: string;
  errorRequest?: string;
  errorNote?: string;
  errorName?: string;
  errorAddress?: string;
  errorCity?: string;
  errorPhone?: string;
  errorEmail?: string;
  errorDistrict?: string;
  companyName?: string;
  insurranceType?: string;
  dateTimeHM?: string;
  dateTimeHH?: string;
  dateTimeM?: string;
  errorCompany?: string;
  insuranceBranch?: string;
  errorDislable?: string;
  errorAmount?: string;
  ownerName?: string;
  errorAddressCompany?: string;
  errorAddressChil?: string;
  errorPhoneCompany?: string;
  errorPhoneChil?: string;
  errorEmailChil?: string;
  carCarrier?: string;
  errorContact?: string;
  errorKm?: string;
  repairContent?: string;
  errorStatus?: string;
}